requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Voice cloning & TTS
TTS>=0.22.0  # Coqui TTS (XTTS support)
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
resampy>=0.4.2

# Audio processing
ffmpeg-python>=0.2.0
pysox>=1.4.2
noisereduce>=3.0.0
webrtcvad>=2.0.10

# Data handling
pandas>=2.1.0
datasets>=2.14.0  # HuggingFace datasets
huggingface-hub>=0.19.0
tqdm>=4.66.0

# NLP & text processing
sentencepiece>=0.1.99
phonemizer>=3.2.1
unidecode>=1.3.7
num2words>=0.5.13

# Arabic text processing
pyarabic>=0.6.15
camel-tools>=1.5.2

# Evaluation metrics
jiwer>=3.0.3  # WER calculation
pesq>=0.0.4   # Perceptual quality
pystoi>=0.3.3 # Speech intelligibility

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.15.0

# API & utilities
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7

# Development tools
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
jupyter>=1.0.0
ipywidgets>=8.1.1

# Optional: GPU acceleration
# nvidia-ml-py3>=7.352.0  # Uncomment if using NVIDIA GPU
